refactor: splitting a ChipContianer from SearchBar
